# portfolio.node/src/docs/openapi.yml
openapi: 3.1.0
info:
  title: Portfolio Backend API
  description: |
    Backend API for the portfolio application following Domain Driven Design principles.
    - Frontend project: portfolio.next
    - Backend project: portfolio.node
    - Backend is supported by MongoDB, Qdrant and Ollama
  version: 1.5.0
...
servers:
  - url: /api
    description: API base URL
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth-token
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    DomainError:
      description: Domain-specific error
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
                description: Domain-specific error message
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
                example: "Authentication required"
    ForbiddenError:
      description: Admin access required
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
                example: "Admin access required"
    ValidationError:
      description: Invalid request data
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
                example: "Invalid request data"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
                example: "Resource not found"
  schemas:
    Blog:
      type: object
      description: Blog entry model representing a blog post
      required:
        - title
        - excerpt
        - body
      properties:
        id:
          type: string
          description: Unique identifier for the blog entry
        title:
          type: string
          description: Blog title
          minLength: 1
          maxLength: 200
        excerpt:
          type: string
          description: Short summary of the blog content
          minLength: 1
          maxLength: 500
        body:
          type: string
          description: Main content of the blog in markdown format
          minLength: 1
        isDraft:
          type: boolean
          description: Whether the blog is in draft status
          default: true
        publishAt:
          type: string
          format: date-time
          description: When to publish the blog
        createdAt:
          type: string
          format: date-time
          description: When the blog was created
        updatedAt:
          type: string
          format: date-time
          description: When the blog was last updated
        tags:
          type: array
          description: List of tags associated with the blog
          items:
            type: string
        slug:
          type: string
          description: URL-friendly version of the title
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
...
paths:
...
  /blogs:
    get:
      tags: [Public]
      summary: Get published blogs
      description: Get all published blog entries
      responses:
        200:
          description: List of published blogs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Blog'
        404:
          $ref: '#/components/responses/NotFoundError'
  /blogs/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Public]
      summary: Get published blog by slug
      description: Get a specific published blog entry by its slug
      responses:
        200:
          description: Blog details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blog'
        404:
          $ref: '#/components/responses/NotFoundError'
  /blogs/search:
    get:
      tags: [Public]
      summary: Search published blogs
      description: Search through published blog entries
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: minScore
          in: query
          schema:
            type: number
            default: 0.7
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    score:
                      type: number
                    document:
                      $ref: '#/components/schemas/Blog'
        400:
          $ref: '#/components/responses/ValidationError'
...
  /search:
    post:
      tags: [Public, Search]
      summary: Global search across all entities
      description: Search across all published content (blogs, projects, pages)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                limit:
                  type: integer
                  default: 10
                minScore:
                  type: number
                  default: 0.7
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    score:
                      type: number
                    document:
                      type: object
        400:
          $ref: '#/components/responses/ValidationError'
...
  /admin/blogs:
    get:
      tags: [Blogs]
      summary: Get all blogs
      description: |
        Get all blogs regardless of draft status.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        200:
          description: List of all blogs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Blog'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags: [Blogs]
      summary: Create new blog
      description: |
        Create a new blog entry.
        Defaults to draft status.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Blog'
      responses:
        201:
          description: Blog created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blog'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
...
  /admin/blogs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          description: Blog identifier
    get:
      tags: [Blogs]
      summary: Get blog by ID
      description: |
        Get a specific blog entry regardless of draft status.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        200:
          description: Blog details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blog'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          description: Blog not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/DomainError'
    put:
      tags: [Blogs]
      summary: Update blog
      description: |
        Update a blog entry.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Blog'
      responses:
        200:
          description: Blog updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blog'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          description: Blog not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/DomainError'
    delete:
      tags: [Blogs]
      summary: Delete blog
      description: |
        Permanently delete a blog entry.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        204:
          description: Blog deleted successfully
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          description: Blog not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/DomainError'
...